% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bug.r
\name{bug}
\alias{bug}
\title{A minimal example}
\description{
roxygen cannot deal with some special characters. In my case the
bug emerged working with backticked square brackets.
}
\details{
\code{[} \code{foo[bar]}

[\code{}foo[bar]: R:%60%20%60foo[bar
}
