% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched.r
\name{matched}
\alias{matched}
\title{A minimal example}
\description{
roxygen cannot deal with some special characters. In my case the
bug emerged working with backticked square brackets.
}
\details{
\code{foo[bar]}
}
